Split main file into several smaller ones
