+++ Generator/Generator.csproj
@@ -48,11 +48,15 @@
     <Reference Include="Microsoft.CSharp" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Generator.cs" />
     <Compile Include="GuidUtility.cs" />
     <Compile Include="AssemblyCollection.cs" />
     <Compile Include="Module.cs" />
+    <Compile Include="NameHelpers.cs" />
+    <Compile Include="ParametricInterfaceManager.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TypeHelpers.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
